Section 4: 84-87
Added statue and created a grabber class. The grabber class was added to the camera POV of the actor, and the rotation of it is being obtained. Added a trace collision called "Gabber". Added the library "Engine/World.h"  and 	#include "DrawDebugHelpers.h" to the Grabber.cpp. Created a pointer that shows a distance of 4 meters to where the actor is pointing at