Section 4: 88-90
Changes made to Grabber. Created functions that uses variables as reference. Created the function HasHit that outputs if an object with the channel of grabber is being currently pointed (hit) by the SweepSingleByChannel function that uses the same parameter for the DrawDebugLine plus a radius specified in a variable in the header.