Sec 4: 91-94
Fixed some issues with VS code that did not allow it to compile. Added some input mappings that are configured on C++ on the Grabber CPP called grab and release. On the BP_Player, added an input action that was configured on the project settings then add the new 2 functions there. Added aPhysicsHandle to the BP_Player. Added 	#include "PhysicsEngine/PhysicsHandleComponent.h" to the grabber.cpp