Section 4: 95 - 96
Created Spheres in Grabber cpp and grabber h.  New function in header, UPhysicsHandleComponent* UGrabber::GetPhysicsHandle() const. In cpp, in hashit() function added PhysicsHandle->GrabComponentAtLocationWithRotation, that allows to grab a component with grabber. New Private attribute in header HoldDistance. In the urneal project, physics were enabled in gargoyle and it is movable. Overall, you can now grab the gargoyle or place a blue circle if it is clicking somewhere else