Section 4: 97
Added the function to Grabber.cpp bool UGrabber::GetGrabbableInReach(FHitResult& OutHitResult)const. That would return a bool if there is an object that can be grabbed. Readded the grabber component to the BP_character. Modified the Release() function in Grabber so the object can be released.
Created a new BP_SecretWall. Created a new component called "triggerComponent.cpp/.h" and added to the BP_SecretWall. Added functionalities on the BP_SecretWall eventgraph for objects that react with the trigger area.