Section 4: 101-106
on TriggerComponent added on TickComponent a function that allows to see what object is interacting with the trigger in the wall. Removed that functionality from the BP. Included Mover.h into the TriggerComponent.h. Added a new function to Mover.h, SetShouldMove to change the ShouldMove  variable. On the SecretwallBP, now instead of outputting locked and unlocked, the should move is activated whenever the gargoyle is in the trigger zone.