Section 4 109-111
Grabber.cpp =
Mover.cpp = Modified the should move option in the Tick component. so the Target location is the always the original location, and if shouldmove is true, add the moveoffset to it.
TigrgerComponent.cpp = added the return option is mover is null.  Added the option so when the gargoyle is present in the trigger area, the wall does not move if the tag is correct and it is not "grabbed".
Grabber.cpp = combined some IF statements for cleaner look and and added in Grab()the code of  HitActor->DetachFromActor(FDetachmentTransformRules::KeepWorldTransform);
Dungeon Map: Addded the statue, the pot and the cell wall for the final puzzle of the level. Added respective Mover and Trigger qualities to them, as well as collisions and physicis of each of them.